Ratings controller and ModelMapper

+++ IzvorniKod/DTPomoziMi/src/main/java/NULL/DTPomoziMi/service/impl/RatingServiceImpl.java
@@ -12,9 +12,12 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 
 import NULL.DTPomoziMi.exception.EntityMissingException;
+import NULL.DTPomoziMi.exception.IllegalAccessException;
 import NULL.DTPomoziMi.model.Rating;
+import NULL.DTPomoziMi.model.User;
 import NULL.DTPomoziMi.repository.RatingRepo;
 import NULL.DTPomoziMi.service.RatingService;
+import NULL.DTPomoziMi.util.UserPrincipalGetter;
 import NULL.DTPomoziMi.web.DTO.RatingDTO;
 
 @Service
@@ -26,9 +29,7 @@ public class RatingServiceImpl implements RatingService {
 	private RatingRepo ratingRepo;
 
 	@Override
-	public Page<Rating> findAll(Pageable pageable) {
-		return ratingRepo.findAll(pageable);
-	}
+	public Page<Rating> findAll(Pageable pageable) { return ratingRepo.findAll(pageable); }
 
 	@Override
 	public Rating save(RatingDTO entity) {
@@ -38,41 +39,37 @@ public class RatingServiceImpl implements RatingService {
 	@Override
 	public Iterable<Rating> saveAll(Iterable<RatingDTO> entities) {
 		Stream<RatingDTO> stream = StreamSupport.stream(entities.spliterator(), false);
-		return ratingRepo
-			.saveAll(
-				stream.map(r -> modelMapper.map(r, Rating.class)).collect(Collectors.toList())
-			);
+		return ratingRepo.saveAll(
+			stream.map(r -> modelMapper.map(r, Rating.class)).collect(Collectors.toList())
+		);
 	}
 
 	@Override
 	public Rating fetch(Long id) {
-		return ratingRepo
-			.findById(id).orElseThrow(() -> new EntityMissingException(Rating.class, id));
+		return ratingRepo.findById(id)
+			.orElseThrow(() -> new EntityMissingException(Rating.class, id));
 	}
 
 	@Override
-	public Optional<Rating> findById(Long id) {
-		return ratingRepo.findById(id);
-	}
+	public Optional<Rating> findById(Long id) { return ratingRepo.findById(id); }
 
 	@Override
-	public boolean existsById(Long id) {
-		return ratingRepo.existsById(id);
-	}
+	public boolean existsById(Long id) { return ratingRepo.existsById(id); }
 
 	@Override
-	public Iterable<Rating> findAll() {
-		return ratingRepo.findAll();
-	}
+	public Iterable<Rating> findAll() { return ratingRepo.findAll(); }
 
 	@Override
-	public long count() {
-		return ratingRepo.count();
-	}
+	public long count() { return ratingRepo.count(); }
 
 	@Override
 	public Rating deleteById(Long id) {
+		User user = UserPrincipalGetter.getPrincipal().getUser();
 		Rating rating = fetch(id);
+
+		if (!user.getIdUser().equals(rating.getRator().getIdUser()))
+			throw new IllegalAccessException("Only author of the rating can delete the rating!");
+
 		ratingRepo.deleteById(id);
 		return rating;
 	}
@@ -91,6 +88,29 @@ public class RatingServiceImpl implements RatingService {
 		Iterable<Rating> iterable = ratingRepo.findAll();
 		ratingRepo.deleteAll();
 		return iterable;
+	}
+
+	@Override
+	public Rating create(RatingDTO rating) {
+		User user = UserPrincipalGetter.getPrincipal().getUser();
+		Rating r = modelMapper.map(rating, Rating.class);
+		r.setRator(user);
+
+		return ratingRepo.save(r);
+	}
+
+	@Override
+	public Rating update(RatingDTO rating, long ratingId) {
+		if (rating.getIdRating() == null || !rating.getIdRating().equals(rating))
+			throw new IllegalArgumentException("Rating id must be preserved!");
+
+		User user = UserPrincipalGetter.getPrincipal().getUser();
+		if (!user.getIdUser().equals(rating.getRator().getIdUser()))
+			throw new IllegalAccessException("Missing permission to update request!");
+
+		Rating r = fetch(ratingId);
+		modelMapper.map(rating, r);
+		return ratingRepo.save(r);
 	}
 
 }
